Split code into package level
Split a the code according to the guidance.
