+++ src-tauri/src/main.rs
@@ -1,28 +1,7 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
-use tauri::{GlobalShortcutManager, Manager,CustomMenuItem,  SystemTray, SystemTrayEvent, SystemTrayMenu, SystemTrayMenuItem};
-
-#[tauri::command]
-fn update_shortcut(shortcut: String, handle: tauri::AppHandle) {
-    handle
-        .global_shortcut_manager()
-        .unregister_all()
-        .unwrap();
-    
-    let window = handle.get_window("main").unwrap();
-    match handle
-        .global_shortcut_manager()
-        .register(&shortcut, move || {
-            println!("Shortcut triggered successfully");
-            window.unminimize().unwrap();
-            window.set_focus().unwrap();
-            window.emit("activate_input_field", {}).unwrap();
-        })
-        {
-        Ok(_) => println!("Shortcut registered successfully"),
-        Err(err) => eprintln!("Failed to register shortcut: {}", err),
-    }
-}
+use tauri::{Manager,CustomMenuItem,  SystemTray, SystemTrayEvent, SystemTrayMenu};
+mod shortcuts;
 
 fn main() {
     let quit = CustomMenuItem::new("quit".to_string(), "Quit");
@@ -34,14 +13,10 @@ fn main() {
 
     tauri::Builder::default()
         .plugin(tauri_plugin_window_state::Builder::default().build())
-        .invoke_handler(tauri::generate_handler![update_shortcut])
+        .invoke_handler(tauri::generate_handler![shortcuts::update_shortcut])
         .system_tray(system_tray)
         .on_system_tray_event(|app, event| match event {
-            SystemTrayEvent::LeftClick {
-                position: _,
-                size: _,
-                ..
-            } => {
+            SystemTrayEvent::LeftClick {..} => {
                 let window = app.get_window("main").unwrap();
                 window.show().unwrap();
                 window.set_focus().unwrap();
